renamed inception score file
